+++ src/app/model/start-initial-visit-guard/index.ts
@@ -1,1 +1,0 @@
-export * from './start-initial-visit-guard';