+++ src/app/model/start-initial-visit-guard/is-initial-visit-to-public-route.store.ts
@@ -1,9 +1,0 @@
-import { createEvent, createStore } from 'effector';
-
-export const $isInitialVisitToPublicRoute = createStore<boolean | null>(null);
-export const setIsInitialVisitToPublicRoute = createEvent<boolean>();
-
-$isInitialVisitToPublicRoute.on(
-  setIsInitialVisitToPublicRoute,
-  (_, payload) => payload,
-);