+++ src/app/ui/app.tsx
@@ -1,10 +1,7 @@
 import type { FC, ReactNode } from 'react';
 
-import { useUnit } from 'effector-react';
+import { MantineProvider } from '@mantine/core';
 
-import { LoadingOverlay, MantineProvider } from '@mantine/core';
-
-import { $isAppVisible } from '../model';
 import { ErrorBoundary } from './error-boundary';
 import { NotificationManager } from './notification-manager';
 import { theme } from './theme';
@@ -12,13 +9,11 @@ import { theme } from './theme';
 import '@mantine/core/styles.css';
 import './styles.module.css';
 
-export const App: FC<{ routerProvider: ReactNode }> = ({ routerProvider }) => {
-  const isAppVisible = useUnit($isAppVisible);
-
+export const App: FC<{ children: ReactNode }> = ({ children }) => {
   return (
     <MantineProvider defaultColorScheme="dark" theme={theme}>
       <ErrorBoundary>
-        {isAppVisible ? routerProvider : <LoadingOverlay visible={true} />}
+        {children}
         <NotificationManager />
       </ErrorBoundary>
     </MantineProvider>